Create README.md
Stroke is a leading cause of death in the United States. According to the CDC, every 40 seconds, a person suffers a stroke in the United States. 
After gathering patient data, we wish to train different machine learning models to predict which patients are likely to suffer from a stroke. 
In this notebook, we will explore data engineering, visualizations, and model training.